feat: finish addReceipt route

+++ migrations/20210418080842-create-receipt-product.js
@@ -9,7 +9,7 @@ module.exports = {
         type: Sequelize.INTEGER
       },
       price: {
-        type: Sequelize.INTEGER
+        type: Sequelize.FLOAT
       },
       quantity: {
         type: Sequelize.INTEGER
